refresh_token: 整理程式碼
